Add a new maths module
